--- cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/dto/request/internal/DownloadFileComponentRequest.java
@@ -1,7 +1,0 @@
-package com.example.cartpostservice.commissions.controller.dto.request.internal;
-
-public record DownloadFileComponentRequest(
-        String code
-) {
-
-}
