+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/controller/dto/response/CommissionElementReadResponse.java
@@ -1,6 +1,5 @@
 package com.example.cartpostservice.commissions.controller.dto.response;
 
-import com.example.cartpostservice.commissions.controller.dto.response.internal.PresignedUrlComponent;
 import org.hexagon.core.vo.PaymentType;
 import java.time.LocalDate;
 import java.util.List;
