+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/infra/client/internal/FileManagementClient.java
@@ -1,8 +1,8 @@
-package com.example.cartpostservice.commissions.controller.internal;
+package com.example.cartpostservice.commissions.infra.client.internal;
 
-import com.example.cartpostservice.commissions.controller.dto.request.internal.DownloadFileComponentRequest;
-import com.example.cartpostservice.commissions.controller.dto.request.internal.FilesRequestDto;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.DownloadFileComponentResponse;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.request.DownloadFileComponentRequest;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.request.FilesRequestDto;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.DownloadFileComponentResponse;
 import org.hexagon.core.dto.Empty;
 import org.hexagon.core.dto.ResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
