+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/infra/client/internal/dto/request/DownloadFileComponentRequest.java
@@ -1,0 +1,7 @@
+package com.example.cartpostservice.commissions.infra.client.internal.dto.request;
+
+public record DownloadFileComponentRequest(
+        String code
+) {
+
+}
