+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/infra/client/internal/dto/request/FilesRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.cartpostservice.commissions.controller.dto.request.internal;
+package com.example.cartpostservice.commissions.infra.client.internal.dto.request;
 
 import java.util.List;
