+++ cartpost-service/src/main/java/com/example/cartpostservice/commissions/service/CommissionsManagerService.java
@@ -2,21 +2,21 @@ package com.example.cartpostservice.commissions.service;
 
 import com.example.cartpostservice.commissions.controller.dto.request.CommissionCreateRequest;
 import com.example.cartpostservice.commissions.controller.dto.request.CommissionUpdateRequest;
-import com.example.cartpostservice.commissions.controller.dto.request.internal.DownloadFileComponentRequest;
-import com.example.cartpostservice.commissions.controller.dto.request.internal.FilesRequestDto;
-import com.example.cartpostservice.commissions.controller.dto.request.internal.TotalPeopleInfoRequestDto;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.request.DownloadFileComponentRequest;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.request.FilesRequestDto;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.request.TotalPeopleInfoRequestDto;
 import com.example.cartpostservice.commissions.controller.dto.response.CommissionCreateResponse;
 import com.example.cartpostservice.commissions.controller.dto.response.CommissionElementReadResponse;
 import com.example.cartpostservice.commissions.controller.dto.response.CommissionUpdateResponse;
 import com.example.cartpostservice.commissions.controller.dto.response.CommissionReadResponse;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.DownloadFileComponentResponse;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.InternalMemberInfo;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.MemberInfoOutput;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.PeopleInfoResponseDto;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.PresignedUrlComponent;
-import com.example.cartpostservice.commissions.controller.internal.ContractClient;
-import com.example.cartpostservice.commissions.controller.internal.FileManagementClient;
-import com.example.cartpostservice.commissions.controller.internal.MemberClient;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.DownloadFileComponentResponse;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.InternalMemberInfo;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.MemberInfoOutput;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.PeopleInfoResponseDto;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.PresignedUrlComponent;
+import com.example.cartpostservice.commissions.infra.client.internal.ContractClient;
+import com.example.cartpostservice.commissions.infra.client.internal.FileManagementClient;
+import com.example.cartpostservice.commissions.infra.client.internal.MemberClient;
 import com.example.cartpostservice.commissions.service.dto.request.CommissionsServiceCommand;
 import com.example.cartpostservice.commissions.service.dto.request.TagServiceCommand;
 import com.example.cartpostservice.commissions.service.dto.response.CommissionsServiceResult;
