+++ cartpost-service/src/test/java/com/example/cartpostservice/cart/controller/MemberClientTest.java
@@ -1,8 +1,8 @@
 package com.example.cartpostservice.cart.controller;
 
-import com.example.cartpostservice.commissions.controller.dto.response.internal.InternalMemberInfo;
-import com.example.cartpostservice.commissions.controller.dto.response.internal.MemberInfoOutput;
-import com.example.cartpostservice.commissions.controller.internal.MemberClient;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.InternalMemberInfo;
+import com.example.cartpostservice.commissions.infra.client.internal.dto.response.MemberInfoOutput;
+import com.example.cartpostservice.commissions.infra.client.internal.MemberClient;
 import com.example.cartpostservice.common.exception.CustomStatusCode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.hexagon.core.dto.ResponseDto;
